requirements.txt
RPi.GPIO
pytz
pysolar
numpy
AWSIoTPythonSDK
timezonefinder
pymemcache
pyserial
